Move error util to new utils folder
