+++ packages/name-service/src/App/routes/Transfer/index.tsx
@@ -6,11 +6,11 @@ import { useAccount, useError, useSdk } from "../../../service";
 import { printableCoin } from "../../../service/helpers";
 import Center from "../../../theme/layout/Center";
 import Stack from "../../../theme/layout/Stack";
+import { getErrorFromStackTrace } from "../../../utils/errors";
 import BackButton from "../../components/BackButton";
 import Loading from "../../components/Loading";
 import YourAccount from "../../components/YourAccount";
 import { pathContract, pathOperationResult, pathTransfer } from "../../paths";
-import { getErrorFromStackTrace } from "../../utils/errors";
 import { OperationResultState } from "../OperationResult";
 import FormTransferName from "./FormTransferName";
 import "./Transfer.less";
