+++ packages/name-service/src/utils/errors.ts
